CP-17381: enable the stderr trace backend
The stderr tracer is the simplest trace backend.

Tracing can be enabled with the "-trace events=<file>" command line
option to QEMU. Where <file> contains a list of trace events to enable
(which can be found in the top-level "trace-events" file.

We may wish to investigate a better trace backend later on.