CP-17381: HACK: disable loading of option ROMs
All options ROMs are provided as part of the BIOS image loaded by the
toolstack and do not need to be loaded by QEMU.

This fixes migrating from qemu-trad which has not allocated any memory
for QEMU to load the options ROMs into.

This should be enabled with a command line option (or similar) and not
hardcoded in this way.