CP-17135: rtl8139: load RxMulOk counter correctly
If a VMSTATE_STRUCT() is used, the fields within this structure cannot
be dependent on the section version.  This is because the
vmstate_load_state() call for the structure uses the version_id in the
VM state description and not the version of the containing section.

Instead, move the tally counter fields into the main VM state
description so the RxMulOk can be correctly dependent on v5.

Without this, QEMU would read 4 fewer bytes form the rtl8139 section,
moving on to the next section too early.  It would see an VM_EOF
(0x00) marker and stop loading any other sections (in particular, it
would not load the following ide section would would break ide).