CP-19645: Add wrapper script to make start QEMU like qemu-trad
Add a wrapper script to transforming qemu-trad command line options so
they are compatible with QEMU and to make QEMU present the same set of
devices as qemu-trad.

This is somewhat preliminary and not every guest configuration will be
supported (e.g., > 2 GiB memory, or multiple VBDs or VIFS may not
work).

Restoring is also not supported (either from a qemu-trad image or a
QEMU one).