CP-17623: qemu-wrapper: use pc-0.10 machine type
pc-0.10 is more compatible with the machine from qemu-trad.

pc-0.13 and earlier machines set the "rombar=0" option on the VGA
device.  This does disable the loading of the VGA ROM (which is what
we want), but has a side-effect it makes QEMU setup the VRAM has MMIO
(thus trapping all accesses) which makes the guest go like treacle.

Instead, disable the VGA ROM by setting an empty "romfile" option.