CP-20437: Create drives for each disk given to the VM
For each disk given to the VM in xenstore, append a -drive argument.
Limit to a maximum of 4 drives as this is what is supported by qemu-trad
(and the usual case for IDE).
Since the information of whether the disk is a CD-ROM or an HDD is not
available in xenstore, a heuristic is used.

This allows booting from a CD and having multiple disks without
PV drivers.