CP-21767 Don't accidently unplug ourselves if  PCI_CLASS_DEVICE is 0x0100
Due to some wild legacy reasons we have to set 0x0100 as a PCI class ID
for Xen Platform device which is a PCI class ID of SCSI Storage
Controller. During Linux kernel initialization all the PV and emulated
SCSI controllers are unplugged which leads to the accidental unplug of
Xen Platform device itself.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>