CA-256686: Disable lock-key-sync
QEMU supports lock_key_sync (built-in heuristics that attempt to key
the Caps and Num Lock keys synchronized with the host), but this expects
VNCClient to send different keysyms for 'A' and 'a,' when using using
EXT_KEY_EXTENSION. Unfortunately XenCenter doesn't support sending
different keysyms for same key, one for uppercase and a different keysym
for lower case.

Workaround this issue, by disabling lock-key-sync.