Add support for CPU hot plug & unplug
Add some patches which tweak QEMU to do CPU hot plug and unplug in the
same fashion that qemu-trad does it. This is done in a legacy way which
means that the main enabling patch is not for upstream. There are a few
patches which should go upstream that are required to get it working,
plus a small qemu-wrapper change.

To add a vCPU, send a device_add QMP message. E.g.:
device_add id=cpu2 driver=qemu32-i386-cpu socket-id=2 core-id=0 thread-id=0

To remove a vCPU, send a device_del QMP message. E.g.:
device_del id=cpu2