CP-17727: Prune devices and other bits of functionality
Prune pieces of QEMU functionality that are not really meant to be
pruned in a normal setup.  Typically all devices are built in but this
is a waste of resources and provides a greater attack area, so disable
building some devices that are not used on XenServer. Also disable a few
other areas of functionality.

There are still more devices that are not used that should be disabled.
However, most of these require code changes as they are entangled in
other QEMU code. This can be done later.

With this commit series, the executable size of QEMU is reduced by
4.6 MiB and the text+data+bss size is reduced by 8.5 MiB.