Rebase on top of v2.10.0
Apart from purely refreshing patches, there are a few changes:
Remove some patches which are not needed.
Add a patch which removes some wrapper functions due to QEMU assuming
that they are needed based on the version of Xen.
Fix up the QMP JSON markup since the checking is now more stringent.
Update the call to the configure script since there are some new knobs
to tweak.
Update the call to xen_map_cache() as it has a new argument.