CA-271407: Fix display of the GVT-g console
Commit 3ef0c573d37b ("console: fix console resize") changed
qemu_console_resize() so that it always allocates a new surface. This
doesn't work with GVT-g because the surface is allocated specially. When
switching to the GVT-g display, it would allocate a surface backed by
the magic GVT-g buffer then immediately call qemu_console_resize() which
allocates a blank, unused surface. This results in a black screen.
Remove the call to qemu_console_resize() -- it is not needed since it
has just replaced the display surface anyway.