CA-271407: gvt-g: Use the PCI version of stdvga
The existing code has a VGT PCI device in QEMU which claims the 00:02.0
BDF but doesn't actually do anything (emulation is done in the kernel
module).  It then uses an ISA version of stdvga to provide the legacy
emulation. This doesn't actually seem to work (or I couldn't make it
work).  Instead, remove the pointless VGT device and use the PCI version
of stdvga at BDF 00:02.0 to provide legacy emulation. This removes some
code and actually works.