CA-271407, CA-266916: Fix issues with GVT-g emulated VRAM
The VRAM for the emulated VGA that is used before the in-guest GVT-g
driver takes over needs to be located somewhere in a PCI BAR otherwise
hvmloader, Xen, QEMU, and the guest won't know where it is. Put the VRAM
address in BAR0 of the emulated PCH as was done with qemu-trad. QEMU
will then advertise this as the address of the linear framebuffer to the
guest.

This fixes a problem where Windows would fall back to a very basic
graphics mode when the GVT-g driver is not in use as well as various
errors with memory accounting which caused QEMU to crash.