CA-285409: Fix Windows RTC issues
Fix two issues provoked by Windows 10's usage of the RTC:

1) Xen emulates the RTC except for the RTC_CENTURY register. qemu-trad
supports reading and writing that register and the rest of the CMOS but
doesn't actually use it for any date calculations.  QEMU upstream does
emulate it fully and triggers RTC_CHANGE events when it is written which
is a problem since QEMU doesn't have the full RTC state (most of it is
in Xen).  Windows writes RTC_CENTURY when setting the RTC even though it
isn't advertised in the ACPI tables and it hasn't changed.  We can't
easily emulate RTC_CENTURY in Xen for trad-compat VMs since the state
will be in QEMU when upgrading. Instead, just don't emit the event when
in trad-compat mode. Going forwards, we should implement RTC_CENTURY
support in Xen for non-compat VMs and then QEMU won't be used for any
RTC emulation.

2) Windows doesn't use the RTC_SET bit in Register B so whenever it
changes the RTC it results in multiple timeoffset ioreqs as each
register is written. Remove rate limiting of RTC_CHANGE events otherwise
some get lost and the time offset stored by XAPI gets out of sync. This
doesn't lessen security since RTC updates were not rate limited with
qemu-trad (and a guest can consume a lot of dom0 CPU by just writing
random IO ports anyway).