CA-285493: Fix unmap issues with NVIDIA GPU passthrough
Previously with QEMU upstream when passing through a GPU, the
nvidia-gridd service (running in a Linux VM) would fail to restart
because MSI would not be enabled properly. The root cause is due to QEMU
unmapping the BARs when bus mastering is disabled. Presumably the guest
tries to read/write the BAR at this point and it subtly corrupts some
state in a way that is only noticed later. Fix it by using a different
object for memory region hooks.

Fix another issue which causes spurious mapping and unmapping of the
BARs due to initializing the BARs with values from the host. While this
probably isn't related to the original problem, it should still be
fixed.

Finally, remove the patch to enable PCI passthrough permissive mode. I
mistakenly thought that this fixed the issue but it turns out it does
nothing. qemu-trad also emits the same warning.