CA-289906: Fix USB Tablet failure on every even migration during live-upgrade
QEMU upgrade failing on every even migration is caused by a bug in qemu-trad.
It didn't migrate USB device state including its bus address. This causes
the bus address of USB Tablet being re-set every time we migrate and
guest kernel usually tries to recover it by reinitializing the device silently.
At this moment the device gets assigned a new bus address which is
not the default 1 that we imply during upgrade procedure. That causes kernel
to recheck HID descriptors which are different in upstream QEMU and thus
failing Tablet upgrade.

Unfortunately, there is no way to recover the previously assigned address
from the migration stream (as it's obviously absent there). So we can only
continue to rely on buggy qemu-trad behavior - assign 0 address and hope
that OS kernel resets the device correctly. To that end, we need to have
identical set of HID descriptors in QEMU for trad-compat mode.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>