CA-293487: Allocate VRAM in a reserved area
When allocating the VRAM, QEMU allocates it immediately after RAM. This
mostly works fine as there is a gap between the end of RAM and the start
of the MMIO region.  However, for instance when QEMU has 2G and the MMIO
region is 2G, there is no gap and QEMU will allocate the VRAM at 2G which
causes breakage since a BAR may be mapped to the same location.

Force RAM block address for guest memory allocation to always point to
a reserved area in guest address space. This version also accounts for
edge cases which previous version didn't: consider 4GB RAM VM with 256MB
MMIO hole and 256MB PCI BAR to place.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>