CP-29935: Add a command to allow querying whether the VM is migratable
Add a command to allow querying whether the VM is migratable which
allows the toolstack to prevent migrations when it won't work.
Like many QMP commands, it returns nothing on success (if the VM is
migratable). If the VM is not migratable, it returns an error message
indicating which device is preventing migration.