CA-299551: Fix sporadic cpu_physical_memory_snapshot_get_dirty assertion
Fix a rare assertion:
exec.c:1215: cpu_physical_memory_snapshot_get_dirty: Assertion `start + length <= snap->end' failed.

Simplified stack trace:
    assert()
    cpu_physical_memory_snapshot_get_dirty()
    memory_region_snapshot_get_dirty()
    vga_draw_graphic() at vga.c:1681

Inspecting the core dump showed that this happened because the snapshot
size was 0 due to region_start == region_end == 0 in vga_draw_graphic.
This happened because VBE had been disabled by the guest (the guest
would do this briefly when changing modes).

The assertion can be reproduced by running this in the guest:

    #include <sys/io.h>

    #define VBE_INDEX_REG 0x01ce
    #define VBE_DATA_REG 0x01cf
    #define VBE_ENABLED_INDEX 0x4
    #define VBE_DISABLED 0

    int main()
    {
            iopl(3);
            outw_p(VBE_ENABLED_INDEX, VBE_INDEX_REG);
            outw_p(VBE_DISABLED, VBE_DATA_REG);
            return 0;
    }

Fix the issue by pulling in a patch from upstream which changes the
region calculation. Although the patch appears to be fixing a different
issue, it fixes this one too because depth == 0 when VBE is disabled.