CA-343454: Fix unplug of IDE drives
We had to change the way IDE disk are added plugged into QEMU because
we couldn't use force-dma`true anymore on -drive, and instead have to
use it on -device. But the code in QEMU that is supposed to unplug IDE
disk only worked with disk added via -drive, but not via -device. This
is a fix.