CP-38631: xen-hvm: Add reserved guest memory address for "tpm-crb-cmd".
When adding more devices like a "tpm-crb", xen_ram_alloc() would
override the ram address of new memory region to the same exact
location. This result in failures. Like:
    include/exec/ram_addr.h:120: ramblock_ptr: Assertion `offset_in_ramblock(block, offset)' failed.

Fix this by overwriting the provided guest address only for known
memory region that we have a reserved address for. We can use the name
of the memory region for that as it is set before xen_ram_alloc() is
called.