CA-366527: Fix passthrough of multiple different devices
This patch incorrectly uses a static structure to share the emu_mask
state. When there are multiple devices passed through, the device added
2nd overwrites the 1st device's emu_mask causing it to behave
incorrectly.

This patch was originally supposed to resolve an issue with multiple
BAR map/unmaps at guest boot. Fix this issue by dropping the patch - I
am unable to reproduce that issue with/without the patch and the tests
runs from the original CA ticket pass in either case.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>