CA-391031: Reinstate rate limiting of RTC_CHANGE events
RTC_CHANGE events are costly to process in dom0 but easy for a guest to
generate. One test showed 4% CPU usage in a guest creating 400% CPU
usage in dom0.

To resolve this, remove the patch which removed rate limiting of
RTC_CHANGE events. At the same time, introduce a new patch to allow
bursts of RTC_CHANGE events to avoid reintroducing the issue described
in the patch being removed.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>