CA-417654: Fix NVME bug which causes WS2025 install failures
The patch series that added NVME namespace support is buggy because
it treats NVME_VOLATILE_WRITE_CACHE as a per-namespace feature when it
is actually a per-controller feature. This resulted in NULL pointer
dereferences as the namespace is not available at that point.
A workaround was previously added to use the nsid instead but Windows
Server 2025 from Sept 2025 (using drivers storport and stornvme
10.0.26100.5074) passes nsid 0 and objects to the invalid nsid response,
causing installation failures.

Fix this by dropping the buggy workaround, replacing the relevant bits
of the namespace support patch with the working upstream version as well
as backporting a few relevant changes on top.

Thanks to Tu Dinh from Vates for debugging this.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>