CA-420202: xen-hvm: Handle framebuffer relocation
It is possible for a guest to relocate the linear framebuffer by writing
to the BAR. This causes QEMU to report an error and if the VM is
migrated, the destination QEMU hits an assertion.

The problem is that updating the BAR causes the PCI core code to call
the del region callback followed by the add region callback but the
xen-hvm code ignores the del region call and therefore the subsequent
call to xen_relocate_memory() during the add region callback fails
because the source address is not correct.

Handle remove region events for the linear framebuffer to fix this
issue.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>